[Backend] Move game_version handling into its own file
Extracts the game_version/dependency resolution (modManifestDeps,
applyIntegrityGameMeta, enrichVersions) out of versions.go into a dedicated
game_version.go, mirroring the concern-per-file layout (e.g. versions_cache.go).
Also trims the docstrings to the repo's lighter style and drops the now-dead
GameVersion guard inside enrichVersions (the fetch is already skipped when it is
set). No behavior change.
